requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchaudio>=2.0.0
pyannote.audio>=3.1.0
huggingface_hub>=0.16.0
requests>=2.31.0
pathlib
logging